+++ ast_canopy/ast_canopy/__init__.py
@@ -1,0 +1,8 @@
+# SPDX-FileCopyrightText: Copyright (c) 2024 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
+# SPDX-License-Identifier: Apache-2.0
+
+from ast_canopy.api import (
+    parse_declarations_from_ast,
+    parse_declarations_from_source,
+    make_ast_from_source,
+)