+++ numba_extensions/curand_device/src/curand_device/curand_bindings.py
@@ -5,7 +5,7 @@ import os
 from collections import defaultdict
 import logging
 
-from canopy import parse_declarations_from_source
+from ast_canopy import parse_declarations_from_source
 from numbast import (
     bind_cxx_enum,
     bind_cxx_structs,