+++ numbast/src/numbast/function.py
@@ -10,8 +10,8 @@ from numba.cuda import declare_device
 from numba.cuda.cudadecl import register_global, register
 from numba.cuda.cudaimpl import lower
 
-from pylibcanopy import execution_space
-from canopy.decl import Function
+from pylibastcanopy import execution_space
+from ast_canopy.decl import Function
 
 from numbast.types import to_numba_type
 from numbast.utils import (